Refactored grades_conversion_service.rb
