+++ app/models/climbing_route.rb
@@ -44,20 +44,4 @@ class ClimbingRoute < ApplicationRecord
   def pitch
     pitches.first
   end
-
-  private
-
-  def sanitize_grade_and_set_standardized_grade
-    case country.grading_system
-    when 'YDS'
-      sanitized_grade = sanitize_yds_grade(grade)
-      self.standardized_grade = YDS[sanitized_grade]
-    when 'EU'
-      # TODO
-      sanitized_grade = grade
-    end
-
-    self.grade = sanitized_grade
-    save!
-  end
 end
